Move from windows::devices to windows::devices::uwb.
