+++ windows/devices/uwb/CMakeLists.txt
@@ -2,7 +2,7 @@
 add_library(windevuwb STATIC "")
 
 set(WINDEVUWB_DIR_PUBLIC_INCLUDE ${CMAKE_CURRENT_LIST_DIR}/include)
-set(WINDEVUWB_DIR_PUBLIC_INCLUDE_PREFIX ${WINDEVUWB_DIR_PUBLIC_INCLUDE}/windows/uwb)
+set(WINDEVUWB_DIR_PUBLIC_INCLUDE_PREFIX ${WINDEVUWB_DIR_PUBLIC_INCLUDE}/windows/devices/uwb)
 
 target_sources(windevuwb
     PRIVATE
