+++ windows/devices/uwb/UwbSession.cxx
@@ -5,21 +5,21 @@
 #include <algorithm>
 #include <memory>
 
-#include <windows/uwb/UwbSession.hxx>
+#include <windows/devices/uwb/UwbSession.hxx>
 
-using namespace windows::devices;
+using namespace windows::devices::uwb;
 
-UwbSession::UwbSession(std::weak_ptr<uwb::UwbSessionEventCallbacks> callbacks, wil::unique_hfile handleDriver) :
-    uwb::UwbSession(std::move(callbacks)),
+UwbSession::UwbSession(std::weak_ptr<::uwb::UwbSessionEventCallbacks> callbacks, wil::unique_hfile handleDriver) :
+    ::uwb::UwbSession(std::move(callbacks)),
     m_handleDriver(std::move(handleDriver))
 {}
 
 void
-UwbSession::ConfigureImpl(const uwb::protocol::fira::UwbSessionData& uwbSessionData)
+UwbSession::ConfigureImpl(const ::uwb::protocol::fira::UwbSessionData &uwbSessionData)
 {
     // TODO: collect options set from uwbConfiguration, translate them to
     // UWB_APP_CONFIG_PARAM, and send them to the driver.
-    
+
     // Populate the session initialization command argument.
     UWB_SESSION_INIT sessionInit;
     sessionInit.sessionId = uwbSessionData.sessionId;
@@ -57,7 +57,7 @@ UwbSession::StopRangingImpl()
 }
 
 void
-UwbSession::AddPeerImpl(uwb::UwbMacAddress /* peerMacAddress */)
+    UwbSession::AddPeerImpl(::uwb::UwbMacAddress /* peerMacAddress */)
 {
     // TODO: two main options for updating the UWB-CLX peer list:
     //  1) Every time a peer is added (on-demand)
