+++ windows/devices/uwb/include/windows/devices/uwb/UwbDeviceDriver.hxx
@@ -6,11 +6,7 @@
 
 #include <initguid.h>
 
-namespace windows
-{
-namespace devices
-{
-namespace uwb
+namespace windows::devices::uwb
 {
 // TODO: remove once the driver DDI is exposed through the Windows SDK/WDK.
 #ifndef GUID_UWB_DEVICE_INTERFACE
@@ -20,8 +16,6 @@ DEFINE_GUID(GUID_UWB_DEVICE_INTERFACE, 0xa7424370, 0x45e5, 0x49c3, 0xae, 0x35, 0
 
 const GUID InterfaceClassUwb = GUID_UWB_DEVICE_INTERFACE;
 const std::string InterfaceClassUwbString = "A7424370-45E5-49C3-AE35-42A3C41B34AD";
-} // namespace uwb
-} // namespace devices
-} // namespace windows
+} // namespace windows::devices::uwb
 
 #endif // UWB_DEVICE_DRIVER_HXX
