+++ windows/devices/uwb/include/windows/devices/uwb/UwbSession.hxx
@@ -14,31 +14,29 @@
 #include <uwb/UwbSessionEventCallbacks.hxx>
 #include <uwb/protocols/fira/UwbConfiguration.hxx>
 
-namespace windows
-{
-namespace devices
+namespace windows::devices::uwb
 {
 /**
- * @brief Windows concrete implementation of a UWB session. 
+ * @brief Windows concrete implementation of a UWB session.
  */
 class UwbSession :
-    public uwb::UwbSession
+    public ::uwb::UwbSession
 {
 public:
     /**
      * @brief Construct a new UwbSession object.
-     * 
+     *
      * @param callbacks The event callback instance.
      * @param handleDriver File handle for a UWB-CX driver instance.
      */
-    UwbSession(std::weak_ptr<uwb::UwbSessionEventCallbacks> callbacks, wil::unique_hfile handleDriver);
+    UwbSession(std::weak_ptr<::uwb::UwbSessionEventCallbacks> callbacks, wil::unique_hfile handleDriver);
 
 private:
     /**
      * @brief TODO
      */
     void
-    ConfigureImpl(const uwb::protocol::fira::UwbSessionData& uwbSessionData) override;
+    ConfigureImpl(const ::uwb::protocol::fira::UwbSessionData& uwbSessionData) override;
 
     /**
      * @brief TODO
@@ -53,16 +51,15 @@ private:
     StopRangingImpl() override;
 
     /**
-     * @brief TODO 
+     * @brief TODO
      */
     void
-    AddPeerImpl(uwb::UwbMacAddress peerMacAddress) override;
+    AddPeerImpl(::uwb::UwbMacAddress peerMacAddress) override;
 
 private:
     wil::unique_hfile m_handleDriver;
 };
 
-} // namespace devices
-} // namespace windows
+} // namespace windows::devices::uwb
 
 #endif // WINDOWS_UWB_SESSION_HXX
